Move TestIssue468 into server_test.go
Per-issue test files were a triage habit, not the project's convention.
Fold the test back into the existing server_test.go suite. No code or
assertion changes — the test runs identically.
